feat: 야구 게임 재시작 기능 개발
