+++ README.md
@@ -4,10 +4,10 @@
 - [x] 사용자로부터 서로 다른 3자리 수를 입력받는다.
   - [x] 이 3자리 수가 포맷에 맞지 않으면 예외를 발생시킨다.
   - [x] 이 3자리 수가 각각 1~9사이가 아니면 예외를 발생시킨다.
-- [ ] 게임을 다시 할지 여부를 입력받는다.
-  - [ ] 1이면 게임을 새로 시작한다.
-  - [ ] 2이면 프로그램을 종료한다.
-  - [ ] 그 외의 입력은 예외를 발생시킨다.
+- [x] 게임을 다시 할지 여부를 입력받는다.
+  - [x] 1이면 게임을 새로 시작한다.
+  - [x] 2이면 프로그램을 종료한다.
+  - [x] 그 외의 입력은 예외를 발생시킨다.
 
 ## 출력
 - [x] 게임 시작 문구를 출력한다.
@@ -30,4 +30,4 @@
 
 ## 프로그램 흐름
 - [ ] 입력 예외가 발생하면 그 부분부터 다시 입력받는다.
-- [ ] 게임을 다시하면 처음부터 다시 시작한다.
+- [x] 게임을 다시하면 처음부터 다시 시작한다.
