--- src/main/java/baseball/controller/Controller.java
@@ -1,6 +1,0 @@
-package baseball.controller;
-
-public interface Controller {
-	
-	void run();
-}
