+++ src/main/java/baseball/io/input/InputHandler.java
@@ -22,4 +22,10 @@ public class InputHandler {
 		String input = reader.readLineAsString();
 		return inputParser.parseBaseballNumbers(input);
 	}
+	
+	public boolean getReplayDecision() {
+		writer.write("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.\n");
+		String input = reader.readLineAsString();
+		return inputParser.parseReplayDecsion(input);
+	}
 }
