+++ src/main/java/baseball/io/input/InputParser.java
@@ -1,5 +1,7 @@
 package baseball.io.input;
 
+import baseball.common.exception.BaseballExceptions;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,5 +14,16 @@ public class InputParser {
 			baseballNumbers.add(Integer.parseInt(String.valueOf(ch)));
 		}
 		return baseballNumbers;
+	}
+	
+	public boolean parseReplayDecsion(String input) {
+		// TODO : 예외 처리
+		if (input.equals("1")) {
+			return true;
+		}
+		if (input.equals("2")) {
+			return false;
+		}
+		throw BaseballExceptions.ILLEGAL_ARGUMENT.get();
 	}
 }
